# packages/rest/src/routes/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"RecordId": {
				"type": "string",
				"example": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e"
			},
...
			"ThreadId": {
				"type": "string",
				"example": "ea4e5e69-fc04-465a-90d2-9f8ff78aa71d"
			},
...
			"BasicMessageRole": {
				"enum": [
					"sender",
					"receiver"
				],
				"type": "string"
			},
			"DidCommBasicMessagesRecord": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/RecordId"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"type": {
						"type": "string"
					},
					"connectionId": {
						"$ref": "#/components/schemas/RecordId"
					},
					"role": {
						"$ref": "#/components/schemas/BasicMessageRole"
					},
					"content": {
						"type": "string"
					},
					"sentTime": {
						"type": "string"
					},
					"threadId": {
						"$ref": "#/components/schemas/ThreadId"
					},
					"parentThreadId": {
						"$ref": "#/components/schemas/ThreadId"
					}
				},
				"required": [
					"id",
					"createdAt",
					"type",
					"connectionId",
					"role",
					"content",
					"sentTime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DidCommBasicMessagesSendOptions": {
				"properties": {
					"connectionId": {
						"$ref": "#/components/schemas/RecordId"
					},
					"content": {
						"type": "string"
					},
					"parentThreadId": {
						"$ref": "#/components/schemas/ThreadId"
					}
				},
				"required": [
					"connectionId",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"tenants": {
				"type": "apiKey",
				"name": "x-tenant-id",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "@credo-ts/rest",
		"version": "0.9.5",
		"description": "Rest API for using Credo over HTTP",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {}
	},
	"paths": {
...
		"/didcomm/basic-messages": {
			"get": {
				"operationId": "FindBasicMessagesByQuery",
				"responses": {
					"200": {
						"description": "BasicMessageRecord[]",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DidCommBasicMessagesRecord"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": "74bcf865-1fdc-45b4-b517-9def02dfd25f",
												"createdAt": "2022-08-18T08:38:40.216Z",
												"type": "BasicMessageRecord",
												"content": "Hello!",
												"sentTime": "2022-08-18T08:38:40.216Z",
												"connectionId": "2aecf74c-3073-4f98-9acb-92415d096834",
												"role": "sender"
											}
										]
									}
								}
							}
						}
					}
				},
				"description": "Retrieve basic messages by connection id",
				"tags": [
					"DIDComm Basic Messages"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"description": "Connection identifier",
						"in": "query",
						"name": "connectionId",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					},
					{
						"in": "query",
						"name": "role",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/BasicMessageRole"
						}
					},
					{
						"in": "query",
						"name": "threadId",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ThreadId"
						}
					},
					{
						"in": "query",
						"name": "parentThreadId",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ThreadId"
						}
					}
				]
			},
			"post": {
				"operationId": "SendMessage",
				"responses": {
					"200": {
						"description": "BasicMessageRecord",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"id": "74bcf865-1fdc-45b4-b517-9def02dfd25f",
											"createdAt": "2022-08-18T08:38:40.216Z",
											"type": "BasicMessageRecord",
											"content": "Hello!",
											"sentTime": "2022-08-18T08:38:40.216Z",
											"connectionId": "2aecf74c-3073-4f98-9acb-92415d096834",
											"role": "sender"
										}
									}
								}
							}
						}
					}
				},
				"description": "Send a basic message to a connection",
				"tags": [
					"DIDComm Basic Messages"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DidCommBasicMessagesSendOptions"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}